refactor: move callModel to separate file
