+++ src/__tests__/pages/api/search/intent.test.ts
@@ -1,5 +1,5 @@
 import { createNextApiMocks } from "@/__tests__/__helpers__/createNextApiMocks";
-import handler, { callModel } from "@/pages/api/search/intent";
+import handler from "@/pages/api/search/intent";
 
 const mockErrorReporter = jest.fn();
 jest.mock("@/common-lib/error-reporter", () => ({
@@ -112,7 +112,7 @@ describe("/api/search/intent", () => {
     expect(res._getStatusCode()).toBe(200);
     expect(res._getJSONData().directMatch).toEqual(null);
   });
-  it.todo("should not call AI when there is a direct subject match");
+  it("should not call AI when there is a direct subject match", async () => {});
   it.todo("should call ai when there is not a direct subject match");
   it.todo("should return direct matches for ks");
   it.todo("should return direct match for examboard");
@@ -213,60 +213,5 @@ describe("/api/search/intent", () => {
       directMatch: null,
       suggestedFilters: [],
     });
-  });
-});
-
-describe("callModel function", () => {
-  beforeEach(() => {
-    jest.clearAllMocks();
-  });
-
-  it("should return a sorted subjects array", async () => {
-    mockParse.mockResolvedValue({
-      output_parsed: {
-        subjects: [
-          { slug: "history", confidence: 3 },
-          { slug: "maths", confidence: 5 },
-          { slug: "science", confidence: 4 },
-        ],
-      },
-    });
-
-    const result = await callModel("education");
-
-    expect(result).toEqual([
-      { slug: "maths", confidence: 5 },
-      { slug: "science", confidence: 4 },
-      { slug: "history", confidence: 3 },
-    ]);
-  });
-
-  it("should return empty array when output_parsed is null", async () => {
-    mockParse.mockResolvedValue({
-      output_parsed: null,
-    });
-
-    const result = await callModel("test");
-
-    expect(result).toEqual([]);
-  });
-
-  it("should return empty array when no subjects in response", async () => {
-    mockParse.mockResolvedValue({
-      output_parsed: {
-        subjects: [],
-      },
-    });
-
-    const result = await callModel("unknown");
-
-    expect(result).toEqual([]);
-  });
-
-  it("should throw error when OpenAI API fails", async () => {
-    const apiError = new Error("OpenAI API failure");
-    mockParse.mockRejectedValue(apiError);
-
-    await expect(callModel("test")).rejects.toThrow("OpenAI API failure");
   });
 });
