+++ src/pages/api/search/intent.ts
@@ -1,7 +1,4 @@
-import OpenAI from "openai";
-import { zodTextFormat } from "openai/helpers/zod";
 import type { NextApiHandler } from "next";
-import { z } from "zod";
 
 import {
   intentRequestSchema,
@@ -11,12 +8,10 @@ import errorReporter from "@/common-lib/error-reporter/errorReporter";
 import OakError from "@/errors/OakError";
 import { findPfMatch } from "@/context/Search/suggestions/findPfMatch";
 import { getSuggestedFilters } from "@/context/Search/suggestions/getSuggestedFilters";
-import { buildSearchIntentPrompt } from "@/utils/promptBuilder";
-import { OAK_SUBJECTS } from "@/context/Search/suggestions/oakCurriculumData";
 import getServerConfig from "@/node-lib/getServerConfig";
+import { callModel } from "@/context/Search/ai/callModel";
 
 const reportError = errorReporter("search-intent");
-const client = new OpenAI();
 
 const handler: NextApiHandler = async (req, res) => {
   const aiSearchEnabled = getServerConfig("aiSearchEnabled");
@@ -65,45 +60,5 @@ const handler: NextApiHandler = async (req, res) => {
     return res.status(500).json({ error: JSON.stringify(error) });
   }
 };
-
-export async function callModel(searchTerm: string) {
-  const subjects = OAK_SUBJECTS.map((subject) => subject.slug);
-  const prompt = buildSearchIntentPrompt(searchTerm, subjects);
-
-  const response = await client.responses.parse({
-    model: "gpt-5-nano",
-    input: prompt,
-    text: {
-      format: zodTextFormat(
-        z.object({
-          subjects: z.array(
-            z.object({
-              slug: z.string(),
-              confidence: z.number().min(1).max(5),
-            }),
-          ),
-        }),
-        "subjects",
-      ),
-    },
-    reasoning: {
-      effort: "low",
-    },
-  });
-
-  const parsedResponse = response.output_parsed;
-
-  if (response.error) {
-    throw response.error;
-  }
-
-  const validSubjects =
-    parsedResponse?.subjects?.filter((subject) =>
-      subjects.includes(subject.slug),
-    ) ?? [];
-
-  validSubjects.sort((a, b) => b.confidence - a.confidence);
-  return validSubjects;
-}
 
 export default handler;
